requirements.txt
gdown
pandas
datasets
sentence-transformers
faiss-cpu
google-genai
numpy
python-dotenv
